requirements.txt
python-binance
pytz
requests